Put include for db back..
Tried using my own database to see if the issues were with my code or with conenctni with Sam's db. When submitting a listing w my Maria db it would take a minute to load and then the if statement would not load (would not print the thank you message or you've missed input go back message).